+++ src/my_app/lib/core/domain/entity/value_object.dart
@@ -3,10 +3,11 @@ import 'dart:convert';
 import 'package:crypto/crypto.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:fpdart/fpdart.dart';
-import 'package:fpvalidate/fpvalidate.dart';
+import 'package:trust_but_verify/trust_but_verify.dart';
 import 'package:uuid/uuid.dart';
 import 'package:very_good_core/app/helpers/extensions/sync_validation_ext.dart';
 import 'package:very_good_core/core/domain/entity/failure.dart';
+import 'package:very_good_core/core/domain/entity/typedef.dart';
 
 @immutable
 abstract class ValueObject<T> {
@@ -15,7 +16,7 @@ abstract class ValueObject<T> {
   @override
   int get hashCode => value.hashCode;
 
-  Either<Failure, T> get value;
+  Result<T> get value;
 
   @override
   bool operator ==(Object other) {
@@ -24,7 +25,7 @@ abstract class ValueObject<T> {
     return other is ValueObject<T> && other.value == value;
   }
 
-  Either<Failure, Unit> get validate => value.fold(left, (T r) => right(unit));
+  Result<Unit> get validate => value.fold(left, (T r) => right(unit));
 
   T getValue() => value.fold((Failure failure) => throw Exception(failure.message), identity);
 
@@ -40,64 +41,64 @@ class UniqueId extends ValueObject<String> {
 
   factory UniqueId.fromUniqueString(String uniqueId) => UniqueId._(
     uniqueId
-        .field('uuid')
+        .trust('uuid')
         .isNotEmpty()
         //TODO: disable this for now since we are using mock data
         //.isUuid()
-        .validateEither()
+        .verifyEither()
         .fold((ValidationError error) => left(Failure.validation(error, uniqueId)), right),
   );
 
   @override
-  final Either<Failure, String> value;
+  final Result<String> value;
 }
 
 class Url extends ValueObject<String> {
   factory Url(String input) => Url._(
     input
-        .field('Url')
+        .trust('Url')
         .isNotEmpty()
         .isUrlStrict()
-        .validateEither()
+        .verifyEither()
         .fold((ValidationError error) => left(Failure.validation(error, input)), right),
   );
   const Url._(this.value);
   @override
-  final Either<Failure, String> value;
+  final Result<String> value;
 }
 
 class EmailAddress extends ValueObject<String> {
   factory EmailAddress(String input) => EmailAddress._(
     input
-        .field('email')
+        .trust('email')
         .isNotEmpty()
-        .isEmailStrict()
-        .validateEither()
+        .isEmail()
+        .verifyEither()
         .fold((ValidationError error) => left(Failure.validation(error, input)), right),
   );
   const EmailAddress._(this.value);
   @override
-  final Either<Failure, String> value;
+  final Result<String> value;
 }
 
 class ValueString extends ValueObject<String> {
   factory ValueString(String input, {required String fieldName}) => ValueString._(
     input
-        .field(fieldName)
+        .trust(fieldName)
         .isNotNull()
         .isNotEmpty()
-        .validateEither()
+        .verifyEither()
         .fold((ValidationError error) => left(Failure.validation(error, input)), right),
   );
   const ValueString._(this.value);
   @override
-  final Either<Failure, String> value;
+  final Result<String> value;
 }
 
 class ValueNumeric extends ValueObject<num> {
   factory ValueNumeric(num? input, {required String fieldName, bool isInt = true}) => ValueNumeric._(
     input
-        .field(fieldName)
+        .trust(fieldName)
         .isNotNull()
         .isNonNegative()
         .bind((num value) {
@@ -109,22 +110,22 @@ class ValueNumeric extends ValueObject<num> {
           }
           return right(value);
         })
-        .validateEither()
+        .verifyEither()
         .fold((ValidationError error) => left(Failure.validation(error, input.toString())), right),
   );
   const ValueNumeric._(this.value);
   @override
-  final Either<Failure, num> value;
+  final Result<num> value;
 }
 
 class Password extends ValueObject<String> {
   factory Password(String input, {bool isHashed = false}) => Password._(
     input
-        .field('password')
+        .trust('password')
         .isNotEmpty()
         .minLength(6) // min password length
         .maxLength(100) // max password length
-        .validateEither()
+        .verifyEither()
         .fold(
           (ValidationError error) => left(Failure.validation(error, input)),
           // encrypt password
@@ -133,7 +134,7 @@ class Password extends ValueObject<String> {
   );
   const Password._(this.value);
   @override
-  final Either<Failure, String> value;
+  final Result<String> value;
 
   static String _encryptPassword(String password, {bool isHashed = false}) =>
       isHashed ? sha256.convert(utf8.encode(password)).toString() : password;