+++ src/my_app/lib/core/domain/interface/i_user_repository.dart
@@ -1,7 +1,6 @@
-import 'package:fpdart/fpdart.dart';
-import 'package:very_good_core/core/domain/entity/failure.dart';
+import 'package:very_good_core/core/domain/entity/typedef.dart';
 import 'package:very_good_core/core/domain/entity/user.dart';
 
 abstract interface class IUserRepository {
-  Future<Either<Failure, User>> get user;
+  Future<Result<User>> get user;
 }