+++ src/my_app/lib/features/auth/data/dto/refresh_token_request.dto.dart
@@ -1,4 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:very_good_core/core/domain/entity/typedef.dart';
 
 part 'refresh_token_request.dto.freezed.dart';
 part 'refresh_token_request.dto.g.dart';
@@ -9,5 +10,5 @@ sealed class RefreshTokenRequestDTO with _$RefreshTokenRequestDTO {
 
   const RefreshTokenRequestDTO._();
 
-  factory RefreshTokenRequestDTO.fromJson(Map<String, dynamic> json) => _$RefreshTokenRequestDTOFromJson(json);
+  factory RefreshTokenRequestDTO.fromJson(Json json) => _$RefreshTokenRequestDTOFromJson(json);
 }