+++ src/my_app/lib/features/home/domain/interface/i_post_repository.dart
@@ -1,7 +1,6 @@
-import 'package:fpdart/fpdart.dart';
-import 'package:very_good_core/core/domain/entity/failure.dart';
+import 'package:very_good_core/core/domain/entity/typedef.dart';
 import 'package:very_good_core/features/home/domain/entity/post.dart';
 
 abstract interface class IPostRepository {
-  Future<Either<Failure, List<Post>>> getPosts();
+  Future<Result<List<Post>>> getPosts();
 }